servers/auth/requirements.txt
asyncpg==0.29.0
bcrypt==4.1.2
gunicorn==20.0.4
grpcio==1.62.1
grpcio-tools==1.62.1
loguru==0.7.2
passlib==1.7.4
pydantic==2.6.4
pydantic-settings==2.2.1
SQLAlchemy==2.0.28
uvicorn==0.13.1